requirements.txt
# MQTT
paho-mqtt~=1.6.1
opencv-python~=4.8.1.78
numpy~=1.26.2

requests~=2.31.0
PyYAML~=6.0.1

torch~=2.1.1
Pillow~=10.1.0